requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tldextract>=5.0.0

# Optional: LLM (install at least one for analysis)
openai>=1.0.0
# anthropic>=0.18.0
